process/requirements.txt
mysql-connector-python
numpy
pandas
sqlalchemy
matplotlib
